requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2-binary
pydantic-settings
bcrypt
PyJWT
loguru
jinja2
reportlab
